Database, Javascript, Node
Hello! I added the data base for the products only. In addition I already implemented the node server and some of the backend requirements.

Badly needed to fix  and redesign are shop, some parts of the landing page, Login sign up.

We also nee additional databases and pages for the checkout and transaction history, user profile and admin side.